+++ config/src/lib.rs
@@ -39,6 +39,7 @@ mod unterm_schemes;
 mod ssh;
 mod terminal;
 mod tls;
+pub mod ui_tokens;
 mod units;
 mod unix;
 mod version;